Function Based Branch:Update Entire Code From Class Based To Function Based

+++ src/Componets/LoadPage/Load.js
@@ -35,10 +35,8 @@ export default function Load() {
         setAnonymous(true);
 
         auth.signInAnonymously().then(user => {
-            //this.setState({user:user.user,uid:user.user.uid});
             setUser(user.user);
             setUID(user.user.uid);
-            //console.log(user);
             db.collection("NoOfGuests").get().then(nog => {
                 var newCount = nog.docs[0].data().count + 1;
                 var name = "Guest" + newCount;
@@ -47,7 +45,6 @@ export default function Load() {
                 db.collection("GuestUsers").doc(user.user.uid).set({Name: name ,uid: user.user.uid,Deleted:false});
 
                 nog.docs[0].ref.update({count: newCount});
-                //this.setState({userName: name});
 
                 setUsername(name);
 
@@ -68,19 +65,14 @@ export default function Load() {
         });
     }
 
-    // ReqForSignIn = () => {
-    //     //this.setState({isAnonymous:false,user:null,uid:"null"});
-    // }
-
     const LoginUser = () =>  {
         auth.signInWithPopup(provider).then(res => {
-            //his.setState({user :res.user,uid:res.user.uid,userName:res.user.displayName});
             setUser(res.user);
             setUID(res.user.uid);
             setUsername(res.user.displayName);
             db.collection("Users").doc(res.user.uid).get().then(user => {
               if(!user.exists){
-                  db.collection("Users").doc(user.id).set({Name: res.user.displayName,uid: user.id}).then(user => {
+                  db.collection("Users").doc(user.id).set({Name: res.user.displayName,uid: user.id}).then(() => {
                       db.collection("Users").doc(uid).get().then(user=>{
                           db.collection("Tasbihs").get().then(tasbihs => {
 
@@ -90,14 +82,12 @@ export default function Load() {
                             console.log(allTasbihs);
 
                             user.ref.collection("Tasbihs").add({count:0,TasbihID:allTasbihs[randPick].id,Name:allTasbihs[randPick].data().Name,Status:'Running'});
-                            //this.setState({loading:false});
                             setLoading(false);
                           });
                       });
                   });
               }else{
                   console.log("User found");
-                  //this.setState({loading:false});
                   setLoading(false);
               }
             })
@@ -106,21 +96,24 @@ export default function Load() {
 
     const LogOutUser = () => {
         auth.signOut().then(() => {
-            user.delete().then(() =>  {
-                db.collection("GuestUsers").doc(this.state.uid).update({Deleted:true}).then(() =>{
-                    // this.setState({
-                    //     user: null,
-                    //     uid:'null',
-                    //     loading: false,
-                    //     isAnonymous: false
-                    // })
-                    setUser(null);
-                    setUID('null');
-                    setLoading(false);
-                    setAnonymous(false);
-                    console.log('user Removed');
+            if(user.isAnonymous){
+                user.delete().then(() =>  {
+                    db.collection("GuestUsers").doc(uid).update({Deleted:true}).then(() =>{
+                        setUser(null);
+                        setUID('null');
+                        setLoading(false);
+                        setAnonymous(false);
+                        console.log('user Removed');
+                    });
                 });
-              });
+            }
+            else{
+                setUser(null);
+                setUID('null');
+                setLoading(false);
+                setAnonymous(false);
+                console.log('user Logout Successfully');
+            }
         },error => {console.log(error)});
     }
 
@@ -131,9 +124,7 @@ export default function Load() {
                     let unsubscibe = db.collection("GuestUsers").doc(user.uid).onSnapshot(data => {
                        if(data.data()){
                         if(!data.data().Deleted){
-                            user.updateProfile({displayName: data.data().Name});   
-                            //this.setState({user:user,uid:user.uid,isAnonymous:user.isAnonymous,userName:data.data().Name});
-                            //this.setState({loading:false});
+                            user.updateProfile({displayName: data.data().Name});
 
                             setUser(user);
                             setUID(user.uid);
@@ -142,7 +133,6 @@ export default function Load() {
                             setLoading(false);
                         }
                         else{
-                            //this.setState({deleted: true});unsubscibe();
                             setDeleted(true);
                             unsubscibe();
                         }
@@ -152,9 +142,6 @@ export default function Load() {
                 else{
                     db.collection("Users").doc(user.uid).onSnapshot(data => {
                         if(data.data()){
-                            // this.setState({user:user,uid:user.uid,isAnonymous:user.isAnonymous,userName:data.data().Name});
-                            // this.setState({loading:false});
-
                             setUser(user);
                             setUID(user.uid);
                             setAnonymous(user.isAnonymous);
@@ -164,14 +151,11 @@ export default function Load() {
                    });
                 }
             }else{
-             // this.setState({loading:false});
              setLoading(false);
             }
         })
     },[]);
-   
-
-    //render() {        
+        
         let Authentic = user || isAnonymous ? 
         <HomePage   click={LogOutUser}
                     skip={isAnonymous}
@@ -184,7 +168,6 @@ export default function Load() {
                     skip={SkipSignIn} />
 
         return ( loading ? <div className="initialize flex">Loading...</div> : Authentic)
-    //}
 }
